requirements.txt
torch
torchvision
streamlit
pillow
numpy
timm
torch-geometric
